ChromeTranslator/pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "deep-translator>=1.11.4",
    "discord-py>=2.5.2",
]
